Refine cross boundary results, separate mods
