+++ src/defid.cpp
@@ -21,6 +21,7 @@
 #include <util/threadnames.h>
 #include <util/translation.h>
 #include <ain_rs_exports.h>
+#include <ffi/ffihelpers.h>
 #include <functional>
 
 const std::function<std::string(const char*)> G_TRANSLATION_FUN = nullptr;
@@ -58,13 +59,7 @@ static void WaitForShutdown()
 //
 static bool AppInit(int argc, char* argv[])
 {
-    {
-        CrossBoundaryResult result;
-        rs_preinit(result);
-        if (!result.ok) {
-            throw std::runtime_error(result.reason.c_str());
-        }
-    }
+    CrossBoundaryCheckedThrow(ain_rs_preinit(result));
 
     InitInterfaces interfaces;
     interfaces.chain = interfaces::MakeChain();
@@ -127,14 +122,10 @@ static bool AppInit(int argc, char* argv[])
         gArgs.SoftSetBoolArg("-server", true);
         // Set this early so that parameter interactions go to console
         InitLogging();
-        {
-            CrossBoundaryResult result;
-            rs_init_logging(result);
-            if (!result.ok) {
-                LogPrintf(result.reason.c_str());
-                return false;
-            }
-        }
+        
+        auto res = CrossBoundaryChecked(ain_rs_init_logging(result));
+        if (!res) return false;
+
         InitParameterInteraction();
         if (!AppInitBasicSetup())
         {
